+++ src/sep.ira/calculator.app/changelog.md
@@ -2,9 +2,9 @@
 
 All notable changes to this forms application will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-[1.0.9663]: https://www.github.com/atecoder/dn.data.finance
+[1.0.9664]: https://www.github.com/atecoder/dn.data.finance
 
-## [1.0.9663] - 2026-06-16
+## [1.0.9664] - 2026-06-17
 - Add report using the report builder.
 
 ## [1.0.9659] - 2026-06-12