+++ src/sep.ira/calculator.xunits/changelog.md
@@ -2,7 +2,10 @@
 
 All notable changes to this unit tests application will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-[1.0.9663]: https://www.github.com/atecoder/dn.data.finance
+[1.0.9664]: https://www.github.com/atecoder/dn.data.finance
+
+## [1.0.9664] - 2026-06-17
+- Add unit tests for the appreciator calculations.
 
 ## [1.0.9663] - 2026-06-16
 - Use titles from the report builder.